Consolidate dry-run e2e tests into deploy.test.ts and undeploy.test.ts
Moves all tests from deploy.dry-run.test.ts and undeploy.dry-run.test.ts
into their respective parent test files and removes the standalone files.
